backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Vector DB
qdrant-client==1.7.3

# Cache
redis==5.0.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# AI/ML - Lightweight for Ollama
openai==1.7.2
httpx==0.26.0

# Embeddings - Required for RAG (lightweight versions)
sentence-transformers>=2.6.0

# PDF Generation
weasyprint==60.2
reportlab==4.1.0
PyPDF2==3.0.1

# DOCX Generation
python-docx>=1.1.1
docxtpl==0.17.0

# Templates
jinja2==3.1.3

# Utils
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator
